+++ src/compute-engine/numerics/linear-algebra.ts
@@ -1,6 +1,6 @@
 // Calculate the determinant of matrix
 
-// Test: determinant([[1,3,7],[2,-1,4],[5,0,2]]) === -54
+// Test: determinant([[1,3,7],[2,-1,4],[5,0,2]]) === 81
 
 export function determinant(matrix: number[][]): number {
   const n = matrix.length;